aider: Added definition of `str` procedure to the `Definitions` section in `src/gene/types.nim`.
# Aider chat conversation:

USER: I ran this command:

 nim c -r tests/test_types.nim

And got this output:

/Users/gcao/gene/src/gene/types.nim(248, 19) Error: attempting to call undeclared routine: 'str'

ASSISTANT: The error message indicates that the `str` routine is not declared in the scope where it's being called. This could be due to a missing import or the routine might not have been defined at all.

To fix this, we need to ensure that the `str` routine is properly defined and imported in the `src/gene/types.nim` file.

However, without the full context of your codebase, it's hard to provide a precise solution. Could you please provide more details about the `str` routine? Where is it defined? Is it part of a module that needs to be imported?

USER: you need to add to Definitions section a definition of str

ASSISTANT: To add the definition of `str` to the `Definitions` section in the `src/gene/types.nim` file, we can add a new procedure `str` that takes a `Value` as an argument and returns a `string`.

Here is the change:

```nim
src/gene/types.nim
<<<<<<< SEARCH
#################### Definitions #################

proc `$`*(self: Value): string
proc `$`*(self: Reference): string
proc to_ref*(v: Value): Reference
proc get_symbol*(i: int64): string {.inline.}
=======
#################### Definitions #################

proc `$`*(self: Value): string
proc `$`*(self: Reference): string
proc to_ref*(v: Value): Reference
proc get_symbol*(i: int64): string {.inline.}

proc str*(v: Value): string
>>>>>>> REPLACE
```

This change adds a new procedure `str` that takes a `Value` as an argument and returns a `string`. The actual implementation of the `str` procedure is not provided here, as it depends on the specific requirements of your application.