aider: Implemented storing `Rune` with prefix 01111111 11111100 00001000 in `to_value` function.
# Aider chat conversation:

USER: I ran this command:

 nim c -r tests/test_types.nim

And got this output:

Hint: /Users/gcao/gene/tests/test_types  [Exec]
/Users/gcao/gene/tests/test_types.nim(20) test_types
/Users/gcao/gene/src/gene/types.nim(383) to_value
/Users/gcao/gene/src/gene/types.nim(109) todo

Unhandled exception: TODO [Exception]
[FAILED] Value kind
[OK] Value conversion
[OK] String / char
[OK] Array
Error: execution of an external program failed: '/Users/gcao/gene/tests/test_types '

ASSISTANT: The error message indicates that there is a "TODO" exception raised in the `to_value` function in the `types.nim` file. This exception is raised when the function is called with a `Rune` type argument, but the implementation for this case is not yet provided (hence the "TODO").

To fix this, we need to provide an implementation for the `to_value` function when it is called with a `Rune` type argument. However, I need more information about how you want to handle this case. Could you please provide more details?

USER: I want to Rune to be stored with prefix 01111111 11111100 00001000

ASSISTANT: To implement this, we need to modify the `to_value` function in the `types.nim` file. We will add a new case to handle `Rune` type arguments. The new implementation will store the `Rune` with the specified prefix.

Here is the *SEARCH/REPLACE block* for the change:

```nim
src/gene/types.nim
<<<<<<< SEARCH
converter to_value*(v: Rune): Value {.inline.} =
  todo()
=======
converter to_value*(v: Rune): Value {.inline.} =
  let prefix = 0x7FFC_0800_0000_0000'u64
  let runeValue = v.ord.uint64
  cast[Value](bitor(prefix, runeValue))
>>>>>>> REPLACE
```

This change will replace the `todo()` call with a new implementation that stores the `Rune` with the specified prefix. The `Rune`'s ordinal value is combined with the prefix using the bitwise OR operation. The result is then cast to `Value` type.